[Tests] Move rng seed into ci/ folder
