+++ tests/util/test_rng_determinism.py
@@ -7,7 +7,7 @@ import pytest
 
 import numpy as np
 import random
-from rng_seed import seed_from_nodeid
+from finn_ci.rng_seed import seed_from_nodeid
 
 pytestmark = pytest.mark.util
